+++ libs/dot-components/src/lib/components/switch/Switch.tsx
@@ -25,9 +25,6 @@ export interface SwitchProps extends CommonProps {
   size?: SwitchSize;
 }
 
-/**
- * @experimental This component is still in development
- */
 export const DotSwitch = ({
   ariaLabel,
   checked = false,